+++ angular-workspace/projects/example-client-app/src/app/app.module.ts
@@ -8,7 +8,7 @@ import { NimbleTextAreaModule, NimbleTextFieldModule, NimbleNumberFieldModule, N
     NimbleMenuItemModule, NimbleCheckboxModule, NimbleToggleButtonModule, NimbleBreadcrumbModule, NimbleBreadcrumbItemModule,
     NimbleIconAddModule, NimbleSwitchModule, NimbleToolbarModule, NimbleMenuButtonModule, NimbleComboboxModule, NimbleTooltipModule,
     NimbleCardButtonModule, NimbleDialogModule, NimbleRadioGroupModule, NimbleRadioModule, NimbleSpinnerModule,
-    NimbleAnchorModule, NimbleAnchorButtonModule } from '@ni/nimble-angular';
+    NimbleAnchorModule, NimbleAnchorButtonModule, NimbleTableColumnTextModule } from '@ni/nimble-angular';
 import { NimbleTableModule } from 'projects/ni/nimble-angular/src/public-api';
 import { AppComponent } from './app.component';
 import { CustomAppComponent } from './customapp/customapp.component';
@@ -58,11 +58,14 @@ import { HeaderComponent } from './header/header.component';
         NimbleAnchorModule,
         NimbleAnchorButtonModule,
         NimbleTableModule,
-        RouterModule.forRoot([
-            { path: '', redirectTo: '/customapp', pathMatch: 'full' },
-            { path: 'customapp', component: CustomAppComponent }
-        ],
-        { useHash: true })
+        NimbleTableColumnTextModule,
+        RouterModule.forRoot(
+            [
+                { path: '', redirectTo: '/customapp', pathMatch: 'full' },
+                { path: 'customapp', component: CustomAppComponent }
+            ],
+            { useHash: true }
+        )
     ],
     providers: [FormBuilder],
     bootstrap: [AppComponent]