+++ packages/nimble-components/src/anchor-tabs/template.ts
@@ -1,0 +1,18 @@
+import { html, ref, slotted, ViewTemplate } from '@microsoft/fast-element';
+import {
+    endSlotTemplate,
+    FoundationElementTemplate,
+    startSlotTemplate
+} from '@microsoft/fast-foundation';
+import type { AnchorTabs, TabsOptions } from '.';
+
+export const template: FoundationElementTemplate<
+ViewTemplate<AnchorTabs>,
+TabsOptions
+> = (context, definition) => html<AnchorTabs>`
+    ${startSlotTemplate(context, definition)}
+    <div ${ref('tablist')} class="tablist" part="tablist" role="tablist">
+        <slot name="anchortab" ${slotted('tabs')}></slot>
+    </div>
+    ${endSlotTemplate(context, definition)}
+`;