+++ packages/nimble-components/src/wafer-map/index.ts
@@ -16,6 +16,7 @@ import {
 } from './types';
 import { DataManager } from './modules/data-manager';
 import { RenderingModule } from './modules/rendering';
+import { ZoomHandler } from './modules/zoom-handler';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -53,6 +54,7 @@ export class WaferMap extends FoundationElement {
     @attr({
         attribute: 'color-scale-mode'
     })
+    public colorScaleMode: WaferMapColorScaleMode = WaferMapColorScaleMode.linear;
 
     /**
      * @internal
@@ -62,10 +64,12 @@ export class WaferMap extends FoundationElement {
     /**
      * @internal
      */
-    @observable public canvasSideLength: number | undefined;
+    public readonly zoomContainer!: HTMLElement;
 
-    @observable public colorScaleMode: WaferMapColorScaleMode = WaferMapColorScaleMode.linear;
-
+    /**
+     * @internal
+     */
+    @observable public canvasSideLength?: number;
     @observable public highlightedValues: string[] = [];
     @observable public dies: WaferMapDie[] = [];
     @observable public colorScale: WaferMapColorScale = {
@@ -74,23 +78,30 @@ export class WaferMap extends FoundationElement {
     };
 
     private renderQueued = false;
-    private dataManager: DataManager | undefined;
-    private renderer: RenderingModule | undefined;
-    private resizeObserver: ResizeObserver | undefined;
+    private dataManager?: DataManager;
+    private renderer?: RenderingModule;
+    private resizeObserver?: ResizeObserver;
+    private zoomHandler?: ZoomHandler;
     public override connectedCallback(): void {
         super.connectedCallback();
         this.resizeObserver = new ResizeObserver(entries => {
             const entry = entries[0];
-            if (entry === undefined) return;
+            if (entry === undefined) {
+                return;
+            }
             const { height, width } = entry.contentRect;
             this.canvasSideLength = Math.min(height, width);
         });
         this.resizeObserver.observe(this);
+        this.canvas.addEventListener('wheel', event => event.preventDefault(), {
+            passive: false
+        });
         this.queueRender();
     }
 
     public override disconnectedCallback(): void {
         super.disconnectedCallback();
+        this.canvas.removeEventListener('wheel', event => event.preventDefault());
         this.resizeObserver!.unobserve(this);
     }
 
@@ -99,7 +110,12 @@ export class WaferMap extends FoundationElement {
      */
     public render(): void {
         this.renderQueued = false;
-        if (this.canvasSideLength === undefined || this.canvasSideLength === 0) return;
+        if (
+            this.canvasSideLength === undefined
+            || this.canvasSideLength === 0
+        ) {
+            return;
+        }
         this.renderer?.clearCanvas(
             this.canvasSideLength,
             this.canvasSideLength
@@ -116,6 +132,14 @@ export class WaferMap extends FoundationElement {
             this.maxCharacters
         );
         this.renderer = new RenderingModule(this.dataManager, this.canvas);
+        this.zoomHandler = new ZoomHandler(
+            this.canvas,
+            this.zoomContainer,
+            this.dataManager,
+            this.renderer,
+            this.canvasSideLength
+        );
+        this.zoomHandler.attachZoomBehavior();
         this.renderer.drawWafer();
     }
 
@@ -163,6 +187,7 @@ export class WaferMap extends FoundationElement {
             this.canvas.width = this.canvasSideLength;
             this.canvas.height = this.canvasSideLength;
         }
+        this.zoomHandler?.resetTransform();
         this.queueRender();
     }